requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
pydantic-settings==2.7.1
pyyaml==6.0.2
